src/requirements.txt
black
cognitive-complexity
flake8-alphabetize
flake8-annotations
flake8-black
flake8-bugbear
flake8-builtins
flake8-cognitive-complexity
flake8-commas
flake8-comprehensions
flake8-polyfill
flake8-variables-names
flake8
pep8-naming
pyflakes
django
django-rest-framework
psycopg2-binary
python-dotenv
drf_yasg
django-extensions